database update 3.0
added a screen that shows an animation for 3 seconds on startup, while checking if user is logged in or not, if not it directs user to login page else it sends user to home page, a page with only a logout button for now. the logout button signs out user from google services and directs user to login page.